--- a/src/Parlot/Properties/AssemblyInfo.cs
+++ b/src/Parlot/Properties/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[module: System.Runtime.CompilerServices.SkipLocalsInit]