--- a/src/Parlot/SkipLocalsInitAttribute.cs
+++ b/src/Parlot/SkipLocalsInitAttribute.cs
@@ -1,0 +1,14 @@
+namespace System.Runtime.CompilerServices;
+
+[AttributeUsage(
+      AttributeTargets.Module
+    | AttributeTargets.Class
+    | AttributeTargets.Struct
+    | AttributeTargets.Interface
+    | AttributeTargets.Constructor
+    | AttributeTargets.Method
+    | AttributeTargets.Property
+    | AttributeTargets.Event, Inherited = false)]
+internal sealed class SkipLocalsInitAttribute : Attribute
+{
+}