… src/Parlot/Character.Generated.cs
… src/Parlot/HexConverter.cs
… test/Parlot.Tests/CharMaskGeneratorTest.cs