@lahma 2022-11-30
If this lands as a release and consumed by Fluid, might be interesting to update the Fluid benchmarks 😉