@sebastienros 2022-11-30
I just published Fluid knowing that Parlot is on its way, I'll do a separate one with just Parlot updated. I didn't want to delay it. Checking if I need to add a net7.0 target, then will run the benchmarks.