@lahma 2022-11-30
Sound good, I'm not in a hurry, Parlot has been fast enough 🙂